+++ test/ui-test/testSuites/global_shared/scripts/settings_names.py
