+++ test/ui-test/testSuites/suite_messaging/shared/scripts/names.py
@@ -5,3 +5,4 @@ from objectmaphelper import *
 from scripts.onboarding_names import *
 from scripts.login_names import *
 from scripts.settings_names import *
+from chat_names import *
