+++ test/ui-test/testSuites/suite_messaging/suite.conf
@@ -1,0 +1,6 @@
+AUT=nim_status_client
+LANGUAGE=Python
+OBJECTMAPSTYLE=script
+TEST_CASES=tst_ChatFlow tst_groupChat
+VERSION=3
+WRAPPERS=Qt
