+++ test/ui-test/testSuites/suite_messaging/tst_groupChat/test.py
