+++ test/ui-test/testSuites/suite_onboarding/tst_statusLoginPassword/test.py
