+++ test/ui-test/testSuites/suite_status/shared/scripts/names.py
@@ -4,11 +4,10 @@ from objectmaphelper import *
 
 #from scripts.global_names import *
 from scripts.onboarding_names import *
+from scripts.settings_names import *
+from scripts.login_names import *
 
-from sections.chat_names import *
 from sections.community_names import *
 from sections.community_portal_names import *
-from scripts.login_names import *
 from sections.search_names import *
-from sections.settings_names import *
 from sections.wallet_names import *
