+++ test/ui-test/testSuites/suite_status/suite.conf
@@ -5,6 +5,6 @@ HOOK_SUB_PROCESSES=false
 IMPLICITAUTSTART=0
 LANGUAGE=Python
 OBJECTMAPSTYLE=script
-TEST_CASES=tst_statusLoginPassword tst_basicChatFlow tst_wallet tst_communityFlows tst_groupChat tst_transaction tst_settingsMenu tst_userIdentity tst_languageSettings tst_searchFlows tst_communityMessageFlows
+TEST_CASES=tst_statusLoginPassword tst_wallet tst_communityFlows tst_transaction tst_settingsMenu tst_userIdentity tst_languageSettings tst_searchFlows tst_communityMessageFlows
 VERSION=3
 WRAPPERS=Qt
